requirements.txt
aiogram
requests
geopy
python-dotenv